+++ components/global-styles/theme.js
@@ -1,7 +1,6 @@
 const theme = {
     colors: {
         black: '#000000',
-        black30: '#303030',
         black27: '#272727',
         'vida-black': '#303030',
         'vida-gray': '#6B6B6B',