+++ pages/product/maternity.js
@@ -1,6 +1,6 @@
 import Layout from 'components/layout'
 import ProductScreen from 'screens/product'
-import { getProductData } from 'services/product'
+import { getAPI } from 'services/common'
 import { getCookieLocale } from 'utils/global-util'
 export async function getServerSideProps({ req, res, resolvedUrl }) {
     let langId = getCookieLocale(req, res) || ''
@@ -9,7 +9,7 @@ export async function getServerSideProps({ req, res, resolvedUrl }) {
             productSlug.lastIndexOf('/') == productSlug.length - 1
                 ? productSlug.slice(0, productSlug.lastIndexOf('/'))
                 : productSlug
-    const prodDt = await getProductData(productName, langId)
+    const prodDt = await getAPI(`vida-product-${productName}`, langId)
 
     return {
         props: {